+++ test/production/app-dir/use-cache-code-hash/app/client/worker.ts
