+++ test/production/app-dir/use-cache-code-hash/app/foo.tsx
